Refactor web to push share and receive mode logic into their respective files
