+++ onionshare/web/receive_mode.py
@@ -1,8 +1,163 @@
+import os
 import tempfile
 from datetime import datetime
-from flask import Request
+from flask import Request, request, render_template, make_response, flash, redirect
+from werkzeug.utils import secure_filename
 
+from ..common import DownloadsDirErrorCannotCreate, DownloadsDirErrorNotWritable
 from .. import strings
+
+
+def receive_routes(web):
+    """
+    The web app routes for receiving files
+    """
+    def index_logic():
+        web.add_request(web.REQUEST_LOAD, request.path)
+
+        if web.common.settings.get('public_mode'):
+            upload_action = '/upload'
+            close_action = '/close'
+        else:
+            upload_action = '/{}/upload'.format(web.slug)
+            close_action = '/{}/close'.format(web.slug)
+
+        r = make_response(render_template(
+            'receive.html',
+            upload_action=upload_action,
+            close_action=close_action,
+            receive_allow_receiver_shutdown=web.common.settings.get('receive_allow_receiver_shutdown')))
+        return web.add_security_headers(r)
+
+    @web.app.route("/<slug_candidate>")
+    def index(slug_candidate):
+        web.check_slug_candidate(slug_candidate)
+        return index_logic()
+
+    @web.app.route("/")
+    def index_public():
+        if not web.common.settings.get('public_mode'):
+            return web.error404()
+        return index_logic()
+
+
+    def upload_logic(slug_candidate=''):
+        """
+        Upload files.
+        """
+        # Make sure downloads_dir exists
+        valid = True
+        try:
+            web.common.validate_downloads_dir()
+        except DownloadsDirErrorCannotCreate:
+            web.add_request(web.REQUEST_ERROR_DOWNLOADS_DIR_CANNOT_CREATE, request.path)
+            print(strings._('error_cannot_create_downloads_dir').format(web.common.settings.get('downloads_dir')))
+            valid = False
+        except DownloadsDirErrorNotWritable:
+            web.add_request(web.REQUEST_ERROR_DOWNLOADS_DIR_NOT_WRITABLE, request.path)
+            print(strings._('error_downloads_dir_not_writable').format(web.common.settings.get('downloads_dir')))
+            valid = False
+        if not valid:
+            flash('Error uploading, please inform the OnionShare user', 'error')
+            if web.common.settings.get('public_mode'):
+                return redirect('/')
+            else:
+                return redirect('/{}'.format(slug_candidate))
+
+        files = request.files.getlist('file[]')
+        filenames = []
+        print('')
+        for f in files:
+            if f.filename != '':
+                # Automatically rename the file, if a file of the same name already exists
+                filename = secure_filename(f.filename)
+                filenames.append(filename)
+                local_path = os.path.join(web.common.settings.get('downloads_dir'), filename)
+                if os.path.exists(local_path):
+                    if '.' in filename:
+                        # Add "-i", e.g. change "foo.txt" to "foo-2.txt"
+                        parts = filename.split('.')
+                        name = parts[:-1]
+                        ext = parts[-1]
+
+                        i = 2
+                        valid = False
+                        while not valid:
+                            new_filename = '{}-{}.{}'.format('.'.join(name), i, ext)
+                            local_path = os.path.join(web.common.settings.get('downloads_dir'), new_filename)
+                            if os.path.exists(local_path):
+                                i += 1
+                            else:
+                                valid = True
+                    else:
+                        # If no extension, just add "-i", e.g. change "foo" to "foo-2"
+                        i = 2
+                        valid = False
+                        while not valid:
+                            new_filename = '{}-{}'.format(filename, i)
+                            local_path = os.path.join(web.common.settings.get('downloads_dir'), new_filename)
+                            if os.path.exists(local_path):
+                                i += 1
+                            else:
+                                valid = True
+
+                basename = os.path.basename(local_path)
+                if f.filename != basename:
+                    # Tell the GUI that the file has changed names
+                    web.add_request(web.REQUEST_UPLOAD_FILE_RENAMED, request.path, {
+                        'id': request.upload_id,
+                        'old_filename': f.filename,
+                        'new_filename': basename
+                    })
+
+                web.common.log('Web', 'receive_routes', '/upload, uploaded {}, saving to {}'.format(f.filename, local_path))
+                print(strings._('receive_mode_received_file').format(local_path))
+                f.save(local_path)
+
+        # Note that flash strings are on English, and not translated, on purpose,
+        # to avoid leaking the locale of the OnionShare user
+        if len(filenames) == 0:
+            flash('No files uploaded', 'info')
+        else:
+            for filename in filenames:
+                flash('Sent {}'.format(filename), 'info')
+
+        if web.common.settings.get('public_mode'):
+            return redirect('/')
+        else:
+            return redirect('/{}'.format(slug_candidate))
+
+    @web.app.route("/<slug_candidate>/upload", methods=['POST'])
+    def upload(slug_candidate):
+        web.check_slug_candidate(slug_candidate)
+        return upload_logic(slug_candidate)
+
+    @web.app.route("/upload", methods=['POST'])
+    def upload_public():
+        if not web.common.settings.get('public_mode'):
+            return web.error404()
+        return upload_logic()
+
+
+    def close_logic(slug_candidate=''):
+        if web.common.settings.get('receive_allow_receiver_shutdown'):
+            web.force_shutdown()
+            r = make_response(render_template('closed.html'))
+            web.add_request(web.REQUEST_CLOSE_SERVER, request.path)
+            return web.add_security_headers(r)
+        else:
+            return redirect('/{}'.format(slug_candidate))
+
+    @web.app.route("/<slug_candidate>/close", methods=['POST'])
+    def close(slug_candidate):
+        web.check_slug_candidate(slug_candidate)
+        return close_logic(slug_candidate)
+
+    @web.app.route("/close", methods=['POST'])
+    def close_public():
+        if not web.common.settings.get('public_mode'):
+            return web.error404()
+        return close_logic()
 
 
 class ReceiveModeWSGIMiddleware(object):
